print digit.c test2 for problem with long int

+++ print_digit.c
@@ -9,7 +9,7 @@
 
 int print_digit (va_list dgt)
 {
-	int len = 0;
+	long int len = 0;
 	long int num = va_arg(dgt, int);
 	long int div = 1;
 	long int tmp;
